Moved DemotePHIsPass to separate file
